backend/requirements.txt
django==3.2
djangorestframework==3.12.4
django-filter==2.4.0
djoser==2.2.2
Pillow==9.3.0